requirements.txt
django>=5.0.1,<6.0.0 # 5.0.1 이상 6.0.0 미만
djangorestframework>=3.14.0,<4.0.0 # django restframework
psycopg2-binary>=2.9.0,<3.0.0
# DRF-Spectacular
drf-spectacular>=0.27.1,<0.28.0
# chatting socket connetion
channels>=4.0.0.,<4.0.1